+++ spec/lib/matterhorn/links/relation/belongs_to_spec.rb
@@ -41,7 +41,7 @@ RSpec.describe "Matterhorn::Links::Relation::BelongsTo" do
   let(:link_set)   { Matterhorn::Links::LinkSet.new(article_class.__link_configs, context: article_class, request_env: request_env)}
 
   it "should set relation to type Links::BelongsTo" do
-    expect(set_member).to be_kind_of(Matterhorn::Links::BelongsTo)
+    expect(set_member).to be_kind_of(Matterhorn::Links::Relation::BelongsTo)
   end
 
   let(:request_env) do
